fix: address ranked and post review feedback
